trackinfo framrate 回退 Signed-off-by: x30034819 <xieyijun3@huawei.com>

+++ test/nativedemo/player/player_demo.cpp
@@ -300,7 +300,7 @@ int32_t PlayerDemo::GetVideoTrackInfo()
         int32_t bitrate = -1;
         int32_t width = -1;
         int32_t height = -1;
-        double framerate = -1;
+        int32_t framerate = -1;
         int32_t type = -1;
         int32_t index = -1;
         for (auto iter = videoTrack.begin(); iter != videoTrack.end(); iter++) {
@@ -308,7 +308,7 @@ int32_t PlayerDemo::GetVideoTrackInfo()
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_BITRATE), bitrate);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_WIDTH), width);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_HEIGHT), height);
-            iter->GetDoubleValue(std::string(PlayerKeys::PLAYER_FRAMERATE), framerate);
+            iter->GetIntValue(std::string(PlayerKeys::PLAYER_FRAMERATE), framerate);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_TRACK_TYPE), type);
             iter->GetIntValue(std::string(PlayerKeys::PLAYER_TRACK_INDEX), index);
             cout << "mime: " << mime.c_str() << ", bitrate: " << bitrate << ", width: " << width << ", height: " <<
